remove uibinder stuff
