+++ src/main/java/com/github/jdramaix/Project.gwt.xml
@@ -14,7 +14,4 @@
     <source path='client'/>
 
     <entry-point class="com.github.jdramaix.client.ConverterApp"/>
-
-    <set-configuration-property name="CssResource.enableGss" value="true"/>
-    <set-configuration-property name="CssResource.gssDefaultInUiBinder" value="true"/>
 </module>
